+++ R/data.R
@@ -9,9 +9,8 @@
 
 .map_tuning <- function(x){
   if(x %in% tabr::tunings$id) return(tabr::tunings$value[x == tabr::tunings$id])
-  if(length(grep("[h-zA-Z]", x)) || length(grep("[a-z][a-z]", x)))
-    stop("Invalid `tuning`.", call. = FALSE)
-  .octave_to_tick(x)
+  if(length(grep("[h-zA-Z]", x))) stop("Invalid `tuning`.", call. = FALSE)
+  gsub(" ", "", .octave_to_tick(x))
 }
 
 #' tabr syntax